Reorganised logger, updated most debugs to infos. Fixes #103
