+++ benchcab/utils/repo.py
@@ -105,7 +105,7 @@ class GitRepo(Repo):
             self.logger.debug(f"Reset to commit {self.commit} (hard reset)")
             repo = git.Repo(self.path)
             repo.head.reset(self.commit, working_tree=True)
-        self.logger.debug(f"Successfully checked out {self.branch} - {self.get_revision()}")
+        self.logger.info(f"Successfully checked out {self.branch} - {self.get_revision()}")
 
     def get_revision(self) -> str:
         """Return the latest revision of the source code.
@@ -186,7 +186,7 @@ class SVNRepo(Repo):
 
         self.subprocess_handler.run_cmd(cmd, verbose=verbose)
 
-        self.logger.debug(f"Successfully checked out {self.path.name} - {self.get_revision()}")
+        self.logger.info(f"Successfully checked out {self.path.name} - {self.get_revision()}")
 
     def get_revision(self) -> str:
         """Return the latest revision of the source code.
